requirements.txt
pandas 
pythonnet
python-dotenv
openpyxl
pyarrow
rich